requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
numpy
PyPDF2
easyocr
Pillow
gTTS
requests
python-multipart
